플레이어에 캐릭터 무브 상속, 플레이어용 상태 클래스 추가, 플레이어 스크립트 3가지로 나누기 (베이스, 흡수, 입력)
