--- Assets/Scripts/CharacterMove/MoveStates.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d3040daa528076245bfea030a9fc4236
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
