+++ Assets/Scripts/Player/Player.Absorb.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3942a57fa4e0ad1469635aa7713c64e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
