+++ Assets/Scripts/Player/Player.Input.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fde4d52065d6691449590595b67a5358
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
