+++ Assets/Scripts/Player/PlayerState/PlayerMoveFalling.cs
@@ -1,0 +1,48 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SuraSang
+{
+    public class PlayerMoveFalling : PlayerMoveState
+    {
+        // TODO : 여기서 메달리기 상태로 이어지게 만들면 될 듯
+
+        public PlayerMoveFalling(CharacterMove characterMove) : base(characterMove) { }
+
+        public override void InitializeState()
+        {
+            _player.OnMove = OnMove;
+        }
+
+        public override void UpdateState() { }
+
+        public override void ClearState() { }
+
+        private void OnMove(Vector2 input)
+        {
+            var dir = _player.MoveDir;
+
+            var y = dir.y - _player.Gravity * _player.FallingGravityMultiplier * Time.deltaTime;
+            y = Mathf.Max(y, -_player.GravityLimit);
+
+            dir.y = 0;
+
+            var inputDir = _player.InputToCameraSpace(input) * _player.Speed;
+            dir = Vector3.MoveTowards(dir, inputDir, _player.AirControl * Time.deltaTime);
+
+            dir.y = y;
+
+            _player.MoveDir = dir;
+
+            if (_controller.isGrounded)
+            {
+                _characterMove.ChangeState(new PlayerMoveGrounded(_characterMove));
+            }
+            else if (_player.IsEdgeDetected() && !_controller.isGrounded)
+            {
+                _characterMove.ChangeState(new PlayerMoveHolding(_characterMove));
+            }
+        }
+    }
+}
