+++ Assets/Scripts/Player/PlayerState/PlayerMoveFalling.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04b570dd18861004eac32c4dd9c91775
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
