+++ Assets/Scripts/Player/PlayerState/PlayerMoveGrounded.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 namespace SuraSang
 {
-    public class CharacterMoveGrounded : CharacterMoveState
+    public class PlayerMoveGrounded : PlayerMoveState
     {
-        public CharacterMoveGrounded(CharacterMove characterMove) : base(characterMove) { }
+        public PlayerMoveGrounded(CharacterMove characterMove) : base(characterMove) { }
 
         private float _lastGroundTime;
 
@@ -17,13 +17,13 @@ namespace SuraSang
 
         public override void InitializeState()
         {
-            _characterMove.OnMove = OnMove;
+            _player.OnMove = OnMove;
 
-            _characterMove.SetAction(ButtonActions.Run, OnRun);
+            _player.SetAction(ButtonActions.Run, OnRun);
             //_characterMove.SetAction(ButtonActions.Crouch, OnCrouch);
-            _characterMove.SetAction(ButtonActions.Jump, OnJump);
+            _player.SetAction(ButtonActions.Jump, OnJump);
 
-            _speed = _characterMove.Speed;
+            _speed = _player.Speed;
         }
         
 
@@ -36,9 +36,9 @@ namespace SuraSang
 
             if (!_controller.isGrounded)
             {
-                if (Time.time - _lastGroundTime > _characterMove.CoyoteTime)
+                if (Time.time - _lastGroundTime > _player.CoyoteTime)
                 {
-                    _characterMove.ChangeState(new CharacterMoveFalling(_characterMove));
+                    _characterMove.ChangeState(new PlayerMoveFalling(_characterMove));
                 }
             }
             else
@@ -49,49 +49,49 @@ namespace SuraSang
 
         public override void ClearState()
         {
-            _characterMove.Crouch(false);
+            _player.Crouch(false);
         }
 
 
         private void OnRun(bool isOn)
         {
-            _characterMove.Crouch(false);
-            _speed = isOn ? _characterMove.Speed * _characterMove.RunMultiplier : _characterMove.Speed;
+            _player.Crouch(false);
+            _speed = isOn ? _player.Speed * _player.RunMultiplier : _player.Speed;
         }
 
         private void OnCrouch(bool isOn)
         {
             _isCrouch = isOn;
-            if (_characterMove.Crouch(isOn))
+            if (_player.Crouch(isOn))
             {
                 _isCrouchFailed = true;
                 return;
             }
-            _speed = isOn ? _characterMove.Speed * _characterMove.CrouchMultiplier : _characterMove.Speed;
+            _speed = isOn ? _player.Speed * _player.CrouchMultiplier : _player.Speed;
         }
 
         private void OnJump(bool isOn)
         {
-            if (isOn && !_characterMove.IsHeadblocked())
+            if (isOn && !_player.IsHeadblocked())
             {
-                _characterMove.ChangeState(new CharacterMoveJumping(_characterMove));
+                _characterMove.ChangeState(new PlayerMoveJumping(_characterMove));
             }
         }
 
         private void OnMove(Vector2 input)
         {
-            var dir = _characterMove.InputToCameraSpace(input);
+            var dir = _player.InputToCameraSpace(input);
 
             if (dir != Vector3.zero)
             {
-                _characterMove.LookVector(dir);
+                _player.LookVector(dir);
             }
 
             dir *= _speed;
 
-            dir.y = _controller.isGrounded ? -1 : _characterMove.MoveDir.y - _characterMove.Gravity * Time.deltaTime;
+            dir.y = _controller.isGrounded ? -1 : _player.MoveDir.y - _player.Gravity * Time.deltaTime;
 
-            _characterMove.MoveDir = dir;
+            _player.MoveDir = dir;
         }
     }
 }
