+++ Assets/Scripts/Player/PlayerState/PlayerMoveState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bed660731faae7347bc9d570ed2e9613
+guid: 2d85ee3b30789ef4baaea4ad94b03c8b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
